[fix]경로 검색 시 시간 정보가 틀리게 나오는 버그 해결
